re-split psp now
